Update Whole-game.Rmd
One of my pet peeves in computer programming help files is that often the examples require background knowledge that some people who need that help file don't have.  In your chapter 2 example, your function works on factors.  I have extensive programming experience using R to solve math problems.  I've never used factors.  So in order for me to follow the example, I have to learn something that is of no interest to me.  Why not make the example be a program that anyone could follow because it uses only basic structures.  For example, how about a program that identifies all the prime factors of an integer?